bot/requirements.txt
python-telegram-bot
elasticsearch==8.10.1
confluent-kafka
requests